pyproject.toml
[project]
name = "midas"
version = "0.1.0"
description = "Investment decision support agent for long-term investors"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-google-genai>=2.0",
    "feedparser>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.scripts]
midas = "midas.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/midas"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
